requirements.txt
streamlit==1.26.0
plotly==5.17.0
pandas==1.4.3
folium==0.14.0
haversine==2.8.0
streamlit-folium==0.13.0
Pillow==9.5.0
DateTime==5.2